backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
pytesseract
pdf2image
Pillow
pymupdf
rapidfuzz
